Add milliseconds to time formatting.
I added milliseconds in the format HH:MM:SS.000 which seems to be the
standard after a little internet research.

Since there is not great formatting support in the JavaScript spec
right now, the format of the local timezone is not quite the same as
before. It will display it as the offset from UTC in hours without
any text description. This appears to be a common issue, as moment.js
ended up cutting this part of the date string due to cross-browser
issues.

An example of the date output after applying this commit is:
```
Sat Dec 20 2014 15:20:17.097 (-05:00)
```